Made all inaccuracy floats, and compressed when networking only

+++ engine/client/client.h
@@ -143,7 +143,7 @@ typedef struct
 	int		weapons;
 	float		maxspeed;
 	float		scr_fov;
-	byte 	weaponInaccuracy;
+	float 	weaponInaccuracy;
 
 	// weapon predict stuff
 	int		weaponsequence;
