+++ game_libs/common/player.h
@@ -319,7 +319,7 @@ public:
 	float m_flPlayAftershock;
 	float m_flNextAmmoBurn;// while charging, when to absorb another unit of player's ammo?
 	ScreenOverlays::OverlayId m_iWeaponScreenOverlay;
-	byte m_iWeaponInaccuracy;
+	float m_iWeaponInaccuracy;
 
 	//Player ID
 	void InitStatusBar( void );
